Add a menu for selecting card packs (#20)
* Refactor options screen

Make the initial screen a user sees on page load a generic options
screen, instead of a dedicated timer select screen. This will make it
easier to add card pack selection to the UI.

* Add the UI for card pack selection

* Download card data on start game click instead of page load

This will make it easier to load only the selected card packs.

This commit also shows an error to the user if card data cannot be loaded.

* Only load cards from the selected packs

* Rename expansion pack to reflect official name

* Always run every step of validation

* Fix indentation